uploaded exposure model for BC
I have replaced the Cascadia.csv exposure model with one that covers all of BC. The mis-match is because the scenario earthquake model includes results that extend beyond the Cascadia region.  We use the BC-level exposure model as the input when running any given scenario and set a distance threshold that determines which assets are considered in the model. In this case, the footprint of the dsra results extends beyond the sample exposure datset that I initially provided for the Cascadia region