Remove unnecessary triggers from our GitHub Actions
Previously, the 5.5-hour GitHub workflow was triggered for every single
push and pull request, which is entirely unnecessary when the source
database dump is currently pulled from elsewhsere and is static.

Also, rename the workflow YAML file, and document more clearly what this
workflow does exactly in the header comments.

Thanks to @wkhchow for alerting me to these frequent unnecessary
GitHub Actions runs, which incidentally tends to fail due to race condition.